Refactor: Splitted the redux code.
- kinfin reducer is splitted into configReducer and analysisReducer.
