+++ src/ui/src/app/store/analysis/actionTypes.js
@@ -1,13 +1,3 @@
-export const INIT_ANALYSIS = "INIT_ANALYSIS";
-export const INIT_ANALYSIS_SUCCESS = "INIT_ANALYSIS_SUCCESS";
-export const INIT_ANALYSIS_FAILURE = "INIT_ANALYSIS_FAILURE";
-export const INIT_ANALYSIS_RESET = "INIT_ANALYSIS_RESET";
-
-export const GET_RUN_STATUS = "GET_RUN_STATUS";
-export const GET_RUN_STATUS_SUCCESS = "GET_RUN_STATUS_SUCCESS";
-export const GET_RUN_STATUS_FAILURE = "GET_RUN_STATUS_FAILURE";
-export const GET_RUN_STATUS_RESET = "GET_RUN_STATUS_RESET";
-
 export const GET_RUN_SUMMARY = "GET_RUN_SUMMARY";
 export const GET_RUN_SUMMARY_SUCCESS = "GET_RUN_SUMMARY_SUCCESS";
 export const GET_RUN_SUMMARY_FAILURE = "GET_RUN_SUMMARY_FAILURE";
@@ -51,11 +41,3 @@ export const GET_PLOT = "GET_PLOT";
 export const GET_PLOT_SUCCESS = "GET_PLOT_SUCCESS";
 export const GET_PLOT_FAILURE = "GET_PLOT_FAILURE";
 export const GET_PLOT_RESET = "GET_PLOT_RESET";
-
-export const SET_SELECTED_ATTRIBUTE_TAXONSET =
-  "SET_SELECTED_ATTRIBUTE_TAXONSET";
-
-export const STORE_CONFIG = "STORE_CONFIG";
-export const STORE_CONFIG_RESET = "STORE_CONFIG_RESET";
-export const RENAME_CONFIG = "RENAME_CONFIG";
-export const DELETE_CONFIG = "DELETE_CONFIG";
