+++ src/ui/src/app/store/analysis/reducers.js
@@ -1,12 +1,4 @@
 import {
-  INIT_ANALYSIS,
-  INIT_ANALYSIS_SUCCESS,
-  INIT_ANALYSIS_FAILURE,
-  INIT_ANALYSIS_RESET,
-  GET_RUN_STATUS,
-  GET_RUN_STATUS_SUCCESS,
-  GET_RUN_STATUS_FAILURE,
-  GET_RUN_STATUS_RESET,
   GET_RUN_SUMMARY,
   GET_RUN_SUMMARY_SUCCESS,
   GET_RUN_SUMMARY_FAILURE,
@@ -39,16 +31,9 @@ import {
   GET_PLOT_SUCCESS,
   GET_PLOT_FAILURE,
   GET_PLOT_RESET,
-  SET_SELECTED_ATTRIBUTE_TAXONSET,
-  STORE_CONFIG,
-  STORE_CONFIG_RESET,
-  RENAME_CONFIG,
-  DELETE_CONFIG,
 } from "./actionTypes";
 
 const initialState = {
-  initAnalysis: { data: null, loading: false, error: null },
-  runStatus: { data: null, loading: false, error: null },
   runSummary: { data: null, loading: false, error: null },
   availableAttributesTaxonsets: { data: null, loading: false, error: null },
   countsByTaxon: { data: null, loading: false, error: null },
@@ -61,50 +46,10 @@ const initialState = {
     loading: false,
     error: null,
   },
-  selectedAttributeTaxonset: { attribute: "all", taxonset: "all" },
-  storeConfig: {
-    data: {},
-  },
 };
 
 const analysisReducer = (state = initialState, action) => {
   switch (action.type) {
-    case INIT_ANALYSIS:
-      return {
-        ...state,
-        initAnalysis: { data: null, loading: true, error: null },
-      };
-    case INIT_ANALYSIS_SUCCESS:
-      return {
-        ...state,
-        initAnalysis: { data: action.payload, loading: false, error: null },
-      };
-    case INIT_ANALYSIS_FAILURE:
-      return {
-        ...state,
-        initAnalysis: { data: null, loading: false, error: action.payload },
-      };
-    case INIT_ANALYSIS_RESET:
-      return { ...state, initAnalysis: initialState.initAnalysis };
-
-    case GET_RUN_STATUS:
-      return {
-        ...state,
-        runStatus: { data: null, loading: true, error: null },
-      };
-    case GET_RUN_STATUS_SUCCESS:
-      return {
-        ...state,
-        runStatus: { data: action.payload, loading: false, error: null },
-      };
-    case GET_RUN_STATUS_FAILURE:
-      return {
-        ...state,
-        runStatus: { data: null, loading: false, error: action.payload },
-      };
-    case GET_RUN_STATUS_RESET:
-      return { ...state, runStatus: initialState.runStatus };
-
     case GET_RUN_SUMMARY:
       return {
         ...state,
@@ -122,7 +67,6 @@ const analysisReducer = (state = initialState, action) => {
       };
     case GET_RUN_SUMMARY_RESET:
       return { ...state, runSummary: initialState.runSummary };
-
     case GET_AVAILABLE_ATTRIBUTES_TAXONSETS:
       return {
         ...state,
@@ -155,7 +99,6 @@ const analysisReducer = (state = initialState, action) => {
         ...state,
         availableAttributesTaxonsets: initialState.availableAttributesTaxonsets,
       };
-
     case GET_COUNTS_BY_TAXON:
       return {
         ...state,
@@ -173,7 +116,6 @@ const analysisReducer = (state = initialState, action) => {
       };
     case GET_COUNTS_BY_TAXON_RESET:
       return { ...state, countsByTaxon: initialState.countsByTaxon };
-
     case GET_CLUSTER_SUMMARY:
       return {
         ...state,
@@ -191,7 +133,6 @@ const analysisReducer = (state = initialState, action) => {
       };
     case GET_CLUSTER_SUMMARY_RESET:
       return { ...state, clusterSummary: initialState.clusterSummary };
-
     case GET_ATTRIBUTE_SUMMARY:
       return {
         ...state,
@@ -209,7 +150,6 @@ const analysisReducer = (state = initialState, action) => {
       };
     case GET_ATTRIBUTE_SUMMARY_RESET:
       return { ...state, attributeSummary: initialState.attributeSummary };
-
     case GET_CLUSTER_METRICS:
       return {
         ...state,
@@ -227,7 +167,6 @@ const analysisReducer = (state = initialState, action) => {
       };
     case GET_CLUSTER_METRICS_RESET:
       return { ...state, clusterMetrics: initialState.clusterMetrics };
-
     case GET_PAIRWISE_ANALYSIS:
       return {
         ...state,
@@ -245,7 +184,6 @@ const analysisReducer = (state = initialState, action) => {
       };
     case GET_PAIRWISE_ANALYSIS_RESET:
       return { ...state, pairwiseAnalysis: initialState.pairwiseAnalysis };
-
     case GET_PLOT:
       return { ...state, plot: { data: null, loading: true, error: null } };
     case GET_PLOT_SUCCESS:
@@ -260,70 +198,6 @@ const analysisReducer = (state = initialState, action) => {
       };
     case GET_PLOT_RESET:
       return { ...state, plot: initialState.plot };
-    case SET_SELECTED_ATTRIBUTE_TAXONSET:
-      return {
-        ...state,
-        selectedAttributeTaxonset: {
-          attribute: action.payload.attribute,
-          taxonset: action.payload.taxonset,
-        },
-      };
-
-    case STORE_CONFIG: {
-      const { sessionId, name, config } = action.payload;
-      const existingData = state.storeConfig?.data || {};
-      return {
-        ...state,
-        storeConfig: {
-          ...state.storeConfig,
-          data: {
-            ...existingData,
-            [sessionId]: { sessionId, name, config },
-          },
-        },
-      };
-    }
-    case RENAME_CONFIG: {
-      const { sessionId, newName } = action.payload;
-      const existing = state.storeConfig.data || {};
-      if (!existing[sessionId]) {
-        return state;
-      }
-
-      return {
-        ...state,
-        storeConfig: {
-          ...state.storeConfig,
-          data: {
-            ...existing,
-            [sessionId]: {
-              ...existing[sessionId],
-              name: newName,
-            },
-          },
-        },
-      };
-    }
-    case DELETE_CONFIG: {
-      const sessionId = action.payload;
-      const { [sessionId]: _, ...remaining } = state.storeConfig.data || {};
-      return {
-        ...state,
-        storeConfig: {
-          ...state.storeConfig,
-          data: remaining,
-        },
-      };
-    }
-
-    case STORE_CONFIG_RESET:
-      return {
-        ...state,
-        storeConfig: {
-          data: {},
-        },
-      };
-
     default:
       return state;
   }
