+++ src/ui/src/app/store/analysis/saga.js
@@ -1,8 +1,6 @@
 import { takeEvery, fork, put, all, call } from "redux-saga/effects";
 
 import {
-  INIT_ANALYSIS,
-  GET_RUN_STATUS,
   GET_AVAILABLE_ATTRIBUTES_TAXONSETS,
   GET_COUNTS_BY_TAXON,
   GET_CLUSTER_SUMMARY,
@@ -13,10 +11,6 @@ import {
   GET_RUN_SUMMARY,
 } from "./actionTypes";
 import {
-  initAnalysisSuccess,
-  initAnalysisFailure,
-  getRunStatusSuccess,
-  getRunStatusFailure,
   getPlotSuccess,
   getPlotFailure,
   getPairwiseAnalysisSuccess,
@@ -33,16 +27,13 @@ import {
   getAvailableAttributesTaxonsetsFailure,
   getRunSummaryFailure,
   getRunSummarySuccess,
-  storeConfig,
 } from "./actions";
 import {
   dispatchErrorToast,
   dispatchSuccessToast,
 } from "../../../utilis/tostNotifications";
 import {
-  initAnalysis,
   getAvailableAttributes,
-  getStatus,
   getCountsByTaxon,
   getRunSummary,
   getClusterSummary,
@@ -52,57 +43,6 @@ import {
   getPlot,
 } from "../../services/client";
 
-function* initAnalysisSaga(action) {
-  const { name, config } = action.payload;
-  try {
-    const response = yield call(initAnalysis, config);
-    if (response.status === "success") {
-      yield put(initAnalysisSuccess(response.data));
-      const payloadForIndexDBStorage = {
-        name,
-        config,
-        sessionId: response.data.session_id,
-      };
-      yield put(storeConfig(payloadForIndexDBStorage));
-      yield call(dispatchSuccessToast, "Analysis initialized successfully!");
-    } else {
-      yield put(initAnalysisFailure(response));
-      yield call(
-        dispatchErrorToast,
-        response?.error?.message || "Failed to initialize analysis"
-      );
-    }
-  } catch (err) {
-    yield put(initAnalysisFailure(err));
-    yield call(
-      dispatchErrorToast,
-      err?.response?.data?.error?.message || "Failed to initialize analysis"
-    );
-  }
-}
-
-function* getRunStatusSaga() {
-  try {
-    const response = yield call(getStatus);
-
-    if (response.status === "success") {
-      yield put(getRunStatusSuccess(response.data));
-      yield call(dispatchSuccessToast, "Run status fetched successfully!");
-    } else {
-      yield put(getRunStatusFailure(response));
-      yield call(
-        dispatchErrorToast,
-        response?.error || "Failed to fetch run status"
-      );
-    }
-  } catch (err) {
-    yield put(getRunStatusFailure(err));
-    yield call(
-      dispatchErrorToast,
-      err?.response?.data?.error || "Failed to fetch run status"
-    );
-  }
-}
 function* getAvailableAttributesSaga() {
   try {
     const response = yield call(getAvailableAttributes);
@@ -302,12 +242,6 @@ function* getPlotsSaga() {
     yield call(dispatchErrorToast, "Failed to fetch plots");
   }
 }
-export function* watchInitAnalysisSaga() {
-  yield takeEvery(INIT_ANALYSIS, initAnalysisSaga);
-}
-export function* watchGetRunStatusSaga() {
-  yield takeEvery(GET_RUN_STATUS, getRunStatusSaga);
-}
 
 export function* watchGetPlotsSaga() {
   yield takeEvery(GET_PLOT, getPlotsSaga);
@@ -339,8 +273,6 @@ export function* watchGetAvailableAttributesSaga() {
 
 function* analysisSaga() {
   yield all([
-    fork(watchInitAnalysisSaga),
-    fork(watchGetRunStatusSaga),
     fork(watchGetAvailableAttributesSaga),
     fork(watchGetRunSummarySaga),
     fork(watchGetCountsByTaxonSaga),
