+++ src/ui/src/app/store/config/actionTypes.js
@@ -1,0 +1,17 @@
+export const INIT_ANALYSIS = "INIT_ANALYSIS";
+export const INIT_ANALYSIS_SUCCESS = "INIT_ANALYSIS_SUCCESS";
+export const INIT_ANALYSIS_FAILURE = "INIT_ANALYSIS_FAILURE";
+export const INIT_ANALYSIS_RESET = "INIT_ANALYSIS_RESET";
+
+export const GET_RUN_STATUS = "GET_RUN_STATUS";
+export const GET_RUN_STATUS_SUCCESS = "GET_RUN_STATUS_SUCCESS";
+export const GET_RUN_STATUS_FAILURE = "GET_RUN_STATUS_FAILURE";
+export const GET_RUN_STATUS_RESET = "GET_RUN_STATUS_RESET";
+
+export const SET_SELECTED_ATTRIBUTE_TAXONSET =
+  "SET_SELECTED_ATTRIBUTE_TAXONSET";
+
+export const STORE_CONFIG = "STORE_CONFIG";
+export const STORE_CONFIG_RESET = "STORE_CONFIG_RESET";
+export const RENAME_CONFIG = "RENAME_CONFIG";
+export const DELETE_CONFIG = "DELETE_CONFIG";
