+++ src/ui/src/app/store/reducers.js
@@ -1,8 +1,10 @@
 import { combineReducers } from "redux";
-import analysisReducer from "./kinfin/reducers";
+import analysisReducer from "./analysis/reducers";
+import configReducer from "./config/reducers";
 
 const rootReducer = combineReducers({
   analysis: analysisReducer,
+  config: configReducer,
 });
 
 export default rootReducer;
