+++ src/ui/src/app/store/sagas.js
@@ -1,7 +1,8 @@
 import { all } from "redux-saga/effects";
 
-import analysisSaga from "./kinfin/saga";
+import analysisSaga from "./analysis/saga";
+import configSaga from "./config/saga";
 
 export default function* rootSaga() {
-  yield all([analysisSaga()]);
+  yield all([analysisSaga(), configSaga()]);
 }
